........
  r8870 | jdh2358 | 2011-01-02 11:53:33 -0600 (Sun, 02 Jan 2011) | 1 line
  
  bumped version number to 1.0.1rc
........
  r8876 | jdh2358 | 2011-01-03 14:34:57 -0600 (Mon, 03 Jan 2011) | 1 line
  
  update pytz to 2010o
........
  r8879 | efiring | 2011-01-03 18:45:25 -0600 (Mon, 03 Jan 2011) | 2 lines
  
  apply patch by Paul Ivanov to move twinx second axis offset to the RHS
........
  r8882 | efiring | 2011-01-04 14:24:12 -0600 (Tue, 04 Jan 2011) | 2 lines
  
  Bugfix: invert_xaxis and invert_yaxis now notify observers. Closes 3123226
........
  r8884 | efiring | 2011-01-04 14:52:42 -0600 (Tue, 04 Jan 2011) | 3 lines
  
  Finance: change volume to float; close  3108059
........
  r8886 | mdehoon | 2011-01-05 07:31:24 -0600 (Wed, 05 Jan 2011) | 2 lines
  
  Replace MacOS.WMAvailable, as the MacOS module is not available with 64-bit Pythons, and deprecated for Python 3.
........
  r8889 | jdh2358 | 2011-01-05 09:59:33 -0600 (Wed, 05 Jan 2011) | 1 line
  
  fix rc file defaults issue for docs; python 2.4 compliance for formlayout
........
  r8891 | weathergod | 2011-01-05 11:44:17 -0600 (Wed, 05 Jan 2011) | 3 lines
  
  Applying a similar fix to r8873 which seemed to only have been applied to the development branch.
  This fixes a math domain error when using log scales.
........
  r8892 | jdh2358 | 2011-01-05 12:14:36 -0600 (Wed, 05 Jan 2011) | 1 line
  
  remove unused params from doc rc; add comment on examples.download params
........
  r8893 | jdh2358 | 2011-01-05 16:04:47 -0600 (Wed, 05 Jan 2011) | 1 line
  
  bumpt the rc version num
........
  r8895 | jdh2358 | 2011-01-05 19:26:23 -0600 (Wed, 05 Jan 2011) | 1 line
  
  fix the plot directive
........
  r8897 | jdh2358 | 2011-01-06 07:50:07 -0600 (Thu, 06 Jan 2011) | 1 line
  
  tag 1.0.1 for release
........
  r8899 | jdh2358 | 2011-01-06 13:42:08 -0600 (Thu, 06 Jan 2011) | 1 line
  
  support relative paths in examples.directory
........

svn path=/trunk/matplotlib/; revision=8900

+++ lib/matplotlib/finance.py
@@ -38,7 +38,7 @@ stock_dt = np.dtype([('date', object),
                      ('close', np.float),
                      ('high', np.float),
                      ('low', np.float),
-                     ('volume', np.int),
+                     ('volume', np.float),
                      ('aclose', np.float)])
 
 
@@ -57,7 +57,7 @@ def parse_yahoo_historical(fh, adjusted=True, asobject=False):
       by the adjusted close, regardless of whether you choose adjusted
       = True|False.
 
-    
+
     *asobject*
       If False (default for compatibility with earlier versions)
       return a list of tuples containing
@@ -101,7 +101,7 @@ def parse_yahoo_historical(fh, adjusted=True, asobject=False):
         dt = datetime.date(*[int(val) for val in datestr.split('-')])
         dnum = date2num(dt)
         open, high, low, close =  [float(val) for val in vals[1:5]]
-        volume = int(vals[5])
+        volume = float(vals[5])
         aclose = float(vals[6])
 
         results.append((dt, dt.year, dt.month, dt.day,
